requirements.txt
python-okx
python-dotenv>=1.0.0
ta==0.10.2
pandas>=2.0.0